1.0.7.5

+++ src/main/java/org/satellite/dev/progiple/hotbed/spawners/handlers/SpawnerBreakHandler.java
@@ -1,27 +1,32 @@
 package org.satellite.dev.progiple.hotbed.spawners.handlers;
 
-import eu.decentsoftware.holograms.api.DHAPI;
-import org.bukkit.Location;
+import com.sk89q.worldedit.bukkit.BukkitAdapter;
+import com.sk89q.worldedit.util.Location;
+import com.sk89q.worldguard.LocalPlayer;
+import com.sk89q.worldguard.WorldGuard;
+import com.sk89q.worldguard.bukkit.WorldGuardPlugin;
+import com.sk89q.worldguard.protection.flags.Flags;
+import com.sk89q.worldguard.protection.regions.RegionContainer;
+import com.sk89q.worldguard.protection.regions.RegionQuery;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockState;
 import org.bukkit.block.CreatureSpawner;
-import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.inventory.InventoryCloseEvent;
-import org.novasparkle.lunaspring.Menus.IMenu;
-import org.novasparkle.lunaspring.Menus.MenuManager;
+import org.novasparkle.lunaspring.API.commands.annotations.LunaHandler;
+import org.novasparkle.lunaspring.API.menus.IMenu;
+import org.novasparkle.lunaspring.API.menus.MenuManager;
 import org.satellite.dev.progiple.hotbed.Tools;
 import org.satellite.dev.progiple.hotbed.configs.Config;
 import org.satellite.dev.progiple.hotbed.configs.SpawnerConfig;
 import org.satellite.dev.progiple.hotbed.spawners.menus.HMenu;
 
-import java.util.Objects;
-
+@LunaHandler
 public class SpawnerBreakHandler implements Listener {
     @EventHandler
     public void onBreakSpawner(BlockBreakEvent e) {
@@ -31,45 +36,53 @@ public class SpawnerBreakHandler implements Listener {
         BlockState state = block.getState();
         if (block.getType() != Material.SPAWNER || !(state instanceof CreatureSpawner)) return;
 
-        CreatureSpawner creatureSpawner = (CreatureSpawner) state;
-        EntityType entityType = creatureSpawner.getSpawnedType();
-        if (entityType == null) entityType = EntityType.ZOMBIE;
+        Location weLocation = new Location(BukkitAdapter.adapt(block.getWorld()), block.getX(), block.getY(), block.getZ());
+        org.bukkit.Location location = block.getLocation();
 
-        Location location = block.getLocation();
-        SpawnerConfig spawnerConfig = SpawnerConfig.getSpawnerCfgs().getOrDefault(location, null);
-        if (spawnerConfig != null) {
-            if (!Tools.isOwner(player, spawnerConfig)) {
-                Config.sendMessage(player, "youNotOwner");
-                return;
-            }
+        LocalPlayer localPlayer = WorldGuardPlugin.inst().wrapPlayer(player);
+        RegionContainer container = WorldGuard.getInstance().getPlatform().getRegionContainer();
+        RegionQuery query = container.createQuery();
 
-            ConfigurationSection section = spawnerConfig.getStorageSection();
-            for (String key : section.getKeys(false)) {
-                if (!Objects.requireNonNull(section.getConfigurationSection(key)).getKeys(false).isEmpty()) {
+        if (query.testBuild(weLocation, localPlayer, Flags.BUILD) || query.testBuild(weLocation, localPlayer, Flags.BLOCK_BREAK)) {
+            CreatureSpawner creatureSpawner = (CreatureSpawner) state;
+            EntityType entityType = creatureSpawner.getSpawnedType();
+            if (entityType == null) entityType = EntityType.ZOMBIE;
+
+            SpawnerConfig spawnerConfig = SpawnerConfig.getSpawnerCfgs().getOrDefault(location, null);
+            if (spawnerConfig != null) {
+                if (Tools.isNotOwner(player, spawnerConfig)) {
+                    Config.sendMessage(player, "youNotOwner");
+                    e.setCancelled(true);
+                    return;
+                }
+
+                if (Tools.getStorageSize(spawnerConfig) > 0) {
                     Config.sendMessage(player, "hasLootInside");
                     e.setCancelled(true);
                     return;
                 }
-            }
-        }
 
-        for (IMenu value : MenuManager.getActiveInventories().values()) {
-            if (value instanceof HMenu) {
-                HMenu hMenu = (HMenu) value;
-                if (hMenu.getSpawnerConfig().getLocation().equals(location)) {
-                    hMenu.getPlayer().closeInventory(InventoryCloseEvent.Reason.PLUGIN);
-                    break;
+                e.setExpToDrop(spawnerConfig.getInt("exp"));
+            }
+
+            for (IMenu value : MenuManager.getActiveInventories().values()) {
+                if (value instanceof HMenu hMenu) {
+                    if (hMenu.getSpawnerConfig().getLocation().equals(location)) {
+                        hMenu.getPlayer().closeInventory(InventoryCloseEvent.Reason.PLUGIN);
+                        break;
+                    }
                 }
             }
-        }
-        if (spawnerConfig != null || Tools.getPlayerChance(player) / 100 >= Math.random()) {
-            int spawnerLevel = spawnerConfig == null ? 1 : spawnerConfig.getInt("level");
 
-            location.getWorld().dropItem(location, Tools.getVirtualSpawner(entityType, player.getName(), spawnerLevel));
-            if (spawnerConfig != null) spawnerConfig.delete();
+            if (spawnerConfig != null || Tools.getPlayerChance(player) / 100 >= Math.random()) {
+                int spawnerLevel = spawnerConfig == null ? 1 : spawnerConfig.getInt("level");
 
-            Config.sendMessage(player, "spawnerReceived");
+                location.getWorld().dropItem(location, Tools.getVirtualSpawner(entityType, player.getName(), spawnerLevel));
+                if (spawnerConfig != null) spawnerConfig.delete();
+
+                Config.sendMessage(player, "spawnerReceived");
+            }
+            else Config.sendMessage(player, "spawnerBroken");
         }
-        else Config.sendMessage(player, "spawnerBroken");
     }
 }
